+++ get-custom-field-values.php
@@ -22,35 +22,6 @@
  */
 
 /*
- * TODO:
- * - Add $order arg to c2c_get_custom(), c2c_get_current_custom()
- * - Create hooks to allow disabling shortcode, shortcode builder, and widget support
- * - Use WP_Query when possible
- * - Facilitate conditional output, maybe via c2c_get_custom_if() where text is only output if post
-     has the custom field AND it equals a specified value (or one of an array of possible values)
-     echo c2c_get_custom_if( 'size', array( 'XL', 'XXL' ), 'Sorry, this size is out of stock.' );
- * - Introduce a 'format' shortcode attribute and template tag argument. Defines the output format for each
-     matching custom field, i.e. c2c_get_custom(..., $format = 'Size %key% has %value%' in stock.')
- * - Support specifying $field as array or comma-separated list of custom fields.
- * - Create args array alternative template tag: c2c_custom_field( $field, $args = array() ) so features
-     can be added and multiple arguments don't have to be explicitly provided. Perhaps transition c2c_get_custom()
-     in plugin v4.0 and detect args.
-     function c2c_get_custom( $field, $args = array() ) {
-       if ( ! empty( $args ) && ! is_array( $args ) ) // Old style usage
-         return c2c_old_get_custom( $field, ... ); // Or: $args = c2c_get_custom_args_into_array( ... );
-       // Do new handling here.
-     }
- * - Support retrieving custom fields for one or more specific post_types
-     c2c_get_custom( 'colors', array( 'post_type' => array( 'pants', 'shorts' ) ) )
- * - Support name filters to run against found custom fields
-     c2c_get_custom( 'colors', array( 'filters' => array( 'strtoupper', 'make_clickable' ) ) )
- * - Since it's shifting to args array, might as well support 'echo'
- * - Move shortcode wizard JS into file so it can be enqueued
- * - Handle serialized custom field values
- * - Gutenberg: Adapt shortcode widget to block editor. Wizard should permit adding custom fields inline or as blocks.
- */
-
-/*
 	Copyright (c) 2004-2020 by Scott Reilly (aka coffee2code)
 
 	This program is free software; you can redistribute it and/or