+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # ChangeLog
 
 ## _(in-progress)_
+* New: Add TODO.md and move existing TODO list from top of main plugin file into it
 * Change: Use HTTPS for link to WP SVN repository in bin script for configuring unit tests, and removed commented-out code
 * Change: Note compatibility through WP 5.4+
 * Change: Update links to coffee2code.com to be HTTPS